requirements.txt
opencv-python
Flask
line-bot-sdk
boto3
gunicorn
